Separate unit and system tests
