+++ tests/system/data/chirps_append_subset_0.nc
