+++ tests/system/data/chirps_append_subset_1.nc
