+++ tests/system/data/chirps_append_subset_with_hole.nc
