+++ tests/system/data/chirps_initial_dataset.nc
