+++ tests/unit/inputs/chirps_example_zarr.json
